* commit 'a721c2dc58e08740441b8725820b1ed8f131a62a': (79 commits)
  FE 필요없는 주석 삭제
  서버 배포 에러를 막기 위한 상수 지정
  Add : Socket MessageChannel use taskExecutor
  chore
  이미지 압축 input useCallback 적용
  압축 이미지 없을 때 예외처리
  style 수정
  이미지 압축때 다른 클릭 방지
  주소변경
  토글기능추가
  20 -> 100KB로 압축기준 수정
  20KB 이상일 때 이미지 압축 수행
  Auto stash before merge of "QA" and "origin/QA"
  이미지 사이즈 제한 1mb
  에러핸들링 함수제작
  meta태그 수정하여 접근성 높이기
  메뉴 수정 추가 실패해도 메시지 뜨는 부분 수정
  문자열 팝업 버그해결
  주소변경
  Refactor : Menu price set a limit
  ...

+++ FE/src/routes/admin/AdminMenuManage/AdminMenuManagePresenter.js
@@ -83,7 +83,8 @@ const AdminMenuManagePresenter = ({
         {
           icon: <span className="gridBtn">삭제</span>,
           callback: () => {
-            deleteItem(row.id);
+            const isDelete = window.confirm(`정말로 ${row.name}을/를 삭제하시겠습니까?`);
+            isDelete && deleteItem(row.id);
           },
         },
       ],
@@ -113,7 +114,7 @@ const AdminMenuManagePresenter = ({
         />
       </div>
       {isPopup && (
-        <div className="MenuManagePopupContainer">
+        <div className="MenuManagePopupContainer" onClick={() => setIsPopup(false)}>
           <MenuManagePopup
             menuPopupData={menuPopupData}
             updateItem={updateItem}
